requirement.txt
tensorflow-gpu==2.4.0
tensorboard==2.5.0
seaborn==0.11.1
pandas==1.2.0
matplotlib=3.3.3
scikit-learn=0.23.2